+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   - `resolution_to_x`, `resolution-to-ratio`, `resolution-to-dppx`, `resolution-to-dpi`, `resolution-to-dpcm`
 - added `to-fraction` function for converting a decimal number to a fractional representation (e.g. `1.5` -> `3/2`)
 - added `list-join` function for joining a list into a string
+- added `archetype-meta` function to retrieve values set on `$CONFIG_META` or `Compass.configuration.meta`
 
 ### Major Changes:
 
@@ -49,6 +50,7 @@
 - `styleguide-diff` now takes two `styleguide` results as params
 - removed `testing` configuration, instead check for `defined?(ArchetypeTestHelpers)`
 - removed `$CONFIG_GLYPHS_NAME`, `$CONFIG_GLYPHS_VERSION`, `$CONFIG_GLYPHS_SVG_ID`, `$CONFIG_GLYPHS_BASE_PATH`, `$CONFIG_GLYPHS_EOT`, `$CONFIG_GLYPHS_FILES`, `$CONFIG_GLYPHS_THRESHOLD`, `$CORE_GLYPHS_MAPPINGS`, `$CONFIG_GLYPHS_MAPPINGS` and instead merged these configs into `$CORE_GLYPHS_LIBRARIES` and `$CONFIG_GLYPHS_LIBRARIES`
+- removed all RTL functions
 
 ## 0.0.1.pre.3 (pre-released)